refactor blog
Summary:

Test Plan:
